refactor(web): share one tabbed documents panel between the gallery and run pages
The gallery page and the run-detail Documents tab duplicated the whole
document-list surface (rows, delete confirm, preview modal, download
wiring, loading/error/empty states) and diverged on layout: the gallery
had purpose tabs, the run tab stacked two labeled sections. Extract
DocumentListPanel owning the tab strip, states, rows, and both modals;
each caller keeps only its data shape — the gallery its server-side
purpose filter + pagination accumulator (footer/onDeleted slots), the
run tab a single fetch filtered client-side per tab. The now-unused
groupDocumentsByPurpose helper and run.inputs/run.outputs locale keys
are removed.
