+++ apps/web/src/pages/documents.tsx
@@ -9,27 +9,11 @@
 
 import { useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { toast } from "sonner";
-import { FileText } from "lucide-react";
-import { getErrorMessage } from "@appstrate/core/errors";
 import { Button } from "@appstrate/ui/components/button";
 import { useCurrentApplicationId } from "../hooks/use-current-application";
-import {
-  useDocuments,
-  useDeleteDocument,
-  useDocumentDownload,
-  type DocumentDto,
-} from "../hooks/use-documents";
-import { usePermissions } from "../hooks/use-permissions";
+import { useDocuments, type DocumentDto } from "../hooks/use-documents";
 import { PageHeader } from "../components/page-header";
-import { LoadingState, ErrorState, EmptyState } from "../components/page-states";
-import { DocumentRow } from "../components/document-row";
-import { DocumentPreview } from "../components/document-preview";
-import { ConfirmModal } from "../components/confirm-modal";
-
-type PurposeFilter = "all" | "user_upload" | "agent_output";
-
-const PURPOSE_TABS: PurposeFilter[] = ["all", "agent_output", "user_upload"];
+import { DocumentListPanel, type PurposeFilter } from "../components/document-list-panel";
 
 export function DocumentsPage() {
   // Remount on application switch so the cursor + accumulated pages reset.
@@ -39,15 +23,10 @@ export function DocumentsPage() {
 
 function DocumentsPageContent() {
   const { t } = useTranslation(["documents", "common"]);
-  const { isMember } = usePermissions();
-  const download = useDocumentDownload();
-  const deleteDoc = useDeleteDocument();
 
   const [purpose, setPurpose] = useState<PurposeFilter>("all");
   const [cursor, setCursor] = useState<string | undefined>(undefined);
   const [loadedPages, setLoadedPages] = useState<DocumentDto[]>([]);
-  const [pendingDelete, setPendingDelete] = useState<DocumentDto | null>(null);
-  const [previewDoc, setPreviewDoc] = useState<DocumentDto | null>(null);
 
   const { data, isLoading, error } = useDocuments({
     purpose: purpose === "all" ? undefined : purpose,
@@ -78,24 +57,6 @@ function DocumentsPageContent() {
     setLoadedPages([]);
   };
 
-  const onDelete = isMember ? (doc: DocumentDto) => setPendingDelete(doc) : undefined;
-
-  const confirmDelete = () => {
-    if (!pendingDelete) return;
-    deleteDoc.mutate(
-      { params: { path: { id: pendingDelete.id } } },
-      {
-        onSuccess: () => {
-          toast.success(t("delete.success"));
-          // Drop it from the accumulator so it disappears without a full reset.
-          setLoadedPages((prev) => prev.filter((d) => d.id !== pendingDelete.id));
-          setPendingDelete(null);
-        },
-        onError: (err) => toast.error(getErrorMessage(err)),
-      },
-    );
-  };
-
   return (
     <div className="p-6">
       <PageHeader
@@ -107,39 +68,17 @@ function DocumentsPageContent() {
         ]}
       />
 
-      <div className="mb-4 flex items-center gap-1">
-        {PURPOSE_TABS.map((p) => (
-          <Button
-            key={p}
-            variant={purpose === p ? "secondary" : "ghost"}
-            size="sm"
-            onClick={() => resetPaging(p)}
-          >
-            {t(`filter.${p}`)}
-          </Button>
-        ))}
-      </div>
-
-      {isLoading ? (
-        <LoadingState />
-      ) : error ? (
-        <ErrorState message={getErrorMessage(error)} />
-      ) : documents.length === 0 ? (
-        <EmptyState message={t("page.empty")} hint={t("page.emptyHint")} icon={FileText} />
-      ) : (
-        <div className="flex flex-col gap-2">
-          {documents.map((doc) => (
-            <DocumentRow
-              key={doc.id}
-              doc={doc}
-              onDownload={download}
-              onDelete={onDelete}
-              onPreview={setPreviewDoc}
-              showRunLink
-            />
-          ))}
-
-          {hasMore && (
+      <DocumentListPanel
+        documents={documents}
+        isLoading={isLoading}
+        error={error}
+        purpose={purpose}
+        onPurposeChange={resetPaging}
+        empty={{ message: t("page.empty"), hint: t("page.emptyHint") }}
+        showRunLink
+        onDeleted={(id) => setLoadedPages((prev) => prev.filter((d) => d.id !== id))}
+        footer={
+          hasMore && (
             <Button
               variant="outline"
               className="mt-2"
@@ -153,23 +92,9 @@ function DocumentsPageContent() {
             >
               {t("page.loadMore")}
             </Button>
-          )}
-        </div>
-      )}
-
-      <ConfirmModal
-        open={!!pendingDelete}
-        onClose={() => setPendingDelete(null)}
-        onConfirm={confirmDelete}
-        title={t("delete.title")}
-        description={t("delete.description", { name: pendingDelete?.name ?? "" })}
-        confirmLabel={t("row.delete")}
-        isPending={deleteDoc.isPending}
+          )
+        }
       />
-
-      {previewDoc && (
-        <DocumentPreview doc={previewDoc} open={!!previewDoc} onClose={() => setPreviewDoc(null)} />
-      )}
     </div>
   );
 }
